Finish up RetroAchievements for UWP, re-stub OpenURL
The primary cause was that the UWP http interface was extremely out of date; as a random guess, I tried WinHTTP's, and it works (and works well) under both Xbox and Windows, so it works for me! Also, I removed OpenURL again because WinRT APIs frankly suck ass. Damn you Microsoft